refactor: 클래스명 변경, api 스펙 변경
