+++ src/main/java/com/example/yeobee/core/expense/domain/UserExpense.java
@@ -1,6 +1,6 @@
 package com.example.yeobee.core.expense.domain;
 
-import com.example.yeobee.core.expense.dto.request.Payer;
+import com.example.yeobee.core.expense.dto.common.UserExpenseDto;
 import com.example.yeobee.core.trip.domain.TripUser;
 import jakarta.persistence.*;
 import java.math.BigDecimal;
@@ -30,7 +30,7 @@ public class UserExpense {
     @JoinColumn(name = "expense_id")
     private Expense expense;
 
-    public UserExpense(Payer payer) {
+    public UserExpense(UserExpenseDto payer) {
         amount = payer.amount();
     }
 }
