+++ src/main/java/com/example/yeobee/core/expense/dto/common/ExpensePhotoDto.java
@@ -1,0 +1,10 @@
+package com.example.yeobee.core.expense.dto.common;
+
+import com.example.yeobee.core.expense.domain.ExpensePhoto;
+
+public record ExpensePhotoDto(Long id, String imageUrl) {
+
+    public ExpensePhotoDto(ExpensePhoto expensePhoto) {
+        this(expensePhoto.getId(), expensePhoto.getImageUrl());
+    }
+}
