+++ src/main/java/com/example/yeobee/core/expense/dto/request/ExpenseUpdateRequestDto.java
@@ -3,14 +3,15 @@ package com.example.yeobee.core.expense.dto.request;
 import com.example.yeobee.core.expense.domain.ExpenseCategory;
 import com.example.yeobee.core.expense.domain.ExpenseMethod;
 import com.example.yeobee.core.expense.domain.ExpenseType;
+import com.example.yeobee.core.expense.dto.common.ExpensePhotoDto;
+import com.example.yeobee.core.expense.dto.common.UserExpenseDto;
 import java.math.BigDecimal;
-import java.time.ZonedDateTime;
 import java.util.List;
 
-public record ExpenseUpdateRequestDto(Long tripId, ZonedDateTime payedAt, ExpenseType expenseType, BigDecimal amount,
+public record ExpenseUpdateRequestDto(Long tripId, ExpenseType expenseType, BigDecimal amount,
                                       String currencyCode, ExpenseMethod expenseMethod,
                                       ExpenseCategory expenseCategory,
                                       String name, Long payerId,
-                                      List<Payer> payerList, List<ExpenseImage> imageList) {
+                                      List<UserExpenseDto> payerList, List<ExpensePhotoDto> imageList) {
 
 }
