+++ src/main/java/com/example/yeobee/core/expense/dto/response/ExpenseCreateResponseDto.java
@@ -2,8 +2,8 @@ package com.example.yeobee.core.expense.dto.response;
 
 import com.example.yeobee.core.expense.domain.ExpenseMethod;
 import com.example.yeobee.core.expense.domain.ExpenseType;
-import com.example.yeobee.core.expense.dto.request.ExpenseImage;
-import com.example.yeobee.core.expense.dto.request.Payer;
+import com.example.yeobee.core.expense.dto.common.ExpensePhotoDto;
+import com.example.yeobee.core.expense.dto.common.UserExpenseDto;
 import java.math.BigDecimal;
 import java.time.ZonedDateTime;
 import java.util.List;
@@ -11,6 +11,6 @@ import java.util.List;
 public record ExpenseCreateResponseDto(Long id, Long tripId, ZonedDateTime payedAt, ExpenseType expenseType,
                                        BigDecimal amount,
                                        String currencyCode, ExpenseMethod expenseMethod, String name, Long payerId,
-                                       List<Payer> payerList, List<ExpenseImage> imageList) {
+                                       List<UserExpenseDto> payerList, List<ExpensePhotoDto> imageList) {
 
 }
