+++ src/main/java/com/example/yeobee/core/expense/dto/response/ExpenseUpdateResponseDto.java
@@ -2,15 +2,14 @@ package com.example.yeobee.core.expense.dto.response;
 
 import com.example.yeobee.core.expense.domain.ExpenseMethod;
 import com.example.yeobee.core.expense.domain.ExpenseType;
-import com.example.yeobee.core.expense.dto.request.ExpenseImage;
-import com.example.yeobee.core.expense.dto.request.Payer;
+import com.example.yeobee.core.expense.dto.common.ExpensePhotoDto;
+import com.example.yeobee.core.expense.dto.common.UserExpenseDto;
 import java.math.BigDecimal;
-import java.time.ZonedDateTime;
 import java.util.List;
 
-public record ExpenseUpdateResponseDto(Long id, Long tripId, ZonedDateTime payedAt, ExpenseType expenseType,
+public record ExpenseUpdateResponseDto(Long id, Long tripId, ExpenseType expenseType,
                                        BigDecimal amount,
                                        String currencyCode, ExpenseMethod expenseMethod, String name, Long payerId,
-                                       List<Payer> payerList, List<ExpenseImage> imageList) {
+                                       List<UserExpenseDto> payerList, List<ExpensePhotoDto> imageList) {
 
 }
